Adds two processors to add to the Virtual controller to configure for physical device
Two Unity Input system processors to allow myself to configure to the physical device in the input manager. Fanatec CSL Elite Pedals for instance are 1.0 without being depressed, 0.5 when depressed half way and 0 when maximum pressure is applied to the pedal (Pressure in a Brake pedal, % thorough pedal travel for throttle and clutch). These processors allow me to remap these values to be 0 when not pressed and 1 when fully pressed.